chore: refactor 'jx step report' to allow more kinds of report
