+++ pkg/cmd/step/report/test_data/junit/multiple_reports/create_quickstarts.junit.xml
